ai-interview-app/backend/python/requirements.txt
# Python dependencies for simple resume parsing
pdfplumber==0.11.0
python-docx==1.1.0

# Face recognition dependencies
face_recognition==1.3.0
numpy==1.24.3
Pillow==10.0.0
dlib==19.24.2
cmake==3.27.0
